+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added possibility to set response code for `response_from` messages in reverse proxy (`kafka` and `http_async`). [#321](https://github.com/Accenture/reactive-interaction-gateway/pull/321)
 - Added new version - `v3` - for internal endpoints to support response code in the `/responses` endpoint
 - Added Helm v3 template to the `deployment` folder [#288](https://github.com/Accenture/reactive-interaction-gateway/issues/288)
+- Added rate limiting for maximum number of WS + SSE + Longpolling connections per minute. Configurable via `MAX_CONNECTIONS_PER_MINUTE` env var, by default it's 5000. [#257](https://github.com/Accenture/reactive-interaction-gateway/issues/257)
+- Added basic setup for the Phoenix LiveDashboard. Now it's accessible at `:4010/dashboard`. [#301](https://github.com/Accenture/reactive-interaction-gateway/issues/301)
 
 ### Changed